Element editor : add new handler for arc to resize angle with mouse.
git-svn-id: svn+ssh://svn.tuxfamily.org/svnroot/qet/qet/trunk@4722 bfdf4180-ca20-0410-9c96-a3a8aa849046

+++ sources/editor/graphicspart/abstractpartellipse.h
@@ -73,8 +73,8 @@ class AbstractPartEllipse : public CustomElementGraphicPart
 	protected:
 		QList<QPointF> saved_points_;
 		QRectF m_rect;
-		int m_start_angle;
-		int m_span_angle;
+		qreal m_start_angle;
+		qreal m_span_angle;
 };
 
 #endif // ABSTRACTPARTELLIPSE_H
